Created core functionality
Created it so the user can record a 'clock-in' or a 'clock-out' with the ability to make sure the input is valid. The punch gets recorded in a .txt file with either a 'CLOCK-IN' or 'CLOCK-OUT' and the exact date/time along with a digital signature (currently hardcoded for me) of the user.