Implements Digital Signature feature
The program checks to see if the 'digital-signature.txt' file exists, if not it creates it and prompts the user for their custom digital signature. If the file does exists but is empty, the program proceeds the same way. If the file exists and contains their signature, it reads it in. Then, when the punch is recorded, it is appended on the end in the 'punch-card.txt' file.